redis cache functionality

+++ db/repository/blog.py
@@ -4,7 +4,8 @@ from db.models.blog import Blog
 from datetime import datetime
 
 
-def create_new_blog(blog: CreateBlog,db: Session, author_id: int = 1):
+
+def create_new_blog(blog: CreateBlog,db: Session, author_id: int):
     blog = Blog(**blog.dict(), author_id=author_id)
     db.add(blog)
     db.commit()
@@ -25,7 +26,9 @@ def list_blogs(db: Session):
 def update_blog(id: int, blog: UpdateBlog, author_id: int, db: Session):
     blog_in_db = db.query(Blog).filter(Blog.id == id).first()
     if not blog_in_db:
-        return
+        return {"error": f"Записи с таким id {id} найти не могу"}
+    if not blog_in_db.author_id == author_id:
+        return {"error": "Только автор может править свой труд"}
     blog_in_db.title = blog.title
     blog_in_db.content = blog.content
     blog_in_db.modifed_at = datetime.now()
@@ -37,7 +40,9 @@ def update_blog(id: int, blog: UpdateBlog, author_id: int, db: Session):
 def delete_blog(id: int, author_id: int, db: Session):
     blog_in_db = db.query(Blog).filter(Blog.id == id)
     if not blog_in_db.first():
-        return {"error":f"не могу найти запись с id {id}"}
+        return {"error": f"Записи с таким id {id} найти не могу"}
+    if not blog_in_db.first().author_id == author_id:
+        return {"error": "Только автор может удалить свой труд"}
     blog_in_db.delete()
     db.commit()
     return {"msg":f"удалена запись с id {id}"}
